requirements.txt
opencv-python
pyaudio
ffmpeg-python
python-dotenv
speechrecognition
gtts


